In Progress; paintAvatar & moveAvatar
moveAvatar should work but cannot be tested before paintAvatar works.

Getting a Type Mismatch (Type 'java/lang/Object' (current frame, stack[2]) is not assignable to 'Environment/Coordinate').

Don't know why. Brain hurty. Logging off.
